+++ Packages/ai.muna.muna/Runtime/Beta/Remote/RemotePredictionService.cs
@@ -75,55 +75,55 @@ namespace Muna.Beta.Services {
         internal RemotePredictionService(MunaClient client) => this.client = client;
 
         private async Task<RemoteValue> ToValue(object? value) => value switch { // INCOMPLETE // Image
-            null              => new() { type = Dtype.Null },
-            float           x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Float32, shape = new int[0] },
-            double          x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Float64, shape = new int[0] },
-            sbyte           x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Int8, shape = new int[0] },
-            short           x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Int16, shape = new int[0] },
-            int             x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Int32, shape = new int[0] },
-            long            x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Int64, shape = new int[0] },
-            byte            x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Uint8, shape = new int[0] },
-            ushort          x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Uint16, shape = new int[0] },
-            uint            x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Uint32, shape = new int[0] },
-            ulong           x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Uint64, shape = new int[0] },
-            bool            x => new() { data = await Upload(new [] { x }.ToStream()), type = Dtype.Bool, shape = new int[0] },
-            float[]         x => new() { data = await Upload(x.ToStream()), type = Dtype.Float32, shape = new [] { x.Length } },
-            double[]        x => new() { data = await Upload(x.ToStream()), type = Dtype.Float64, shape = new [] { x.Length } },
-            sbyte[]         x => new() { data = await Upload(x.ToStream()), type = Dtype.Int8, shape = new [] { x.Length } },
-            short[]         x => new() { data = await Upload(x.ToStream()), type = Dtype.Int16, shape = new [] { x.Length } },
-            int[]           x => new() { data = await Upload(x.ToStream()), type = Dtype.Int32, shape = new [] { x.Length } },
-            long[]          x => new() { data = await Upload(x.ToStream()), type = Dtype.Int64, shape = new [] { x.Length } },
-            byte[]          x => new() { data = await Upload(x.ToStream()), type = Dtype.Uint8, shape = new [] { x.Length } },
-            ushort[]        x => new() { data = await Upload(x.ToStream()), type = Dtype.Uint16, shape = new [] { x.Length } },
-            uint[]          x => new() { data = await Upload(x.ToStream()), type = Dtype.Uint32, shape = new [] { x.Length } },
-            ulong[]         x => new() { data = await Upload(x.ToStream()), type = Dtype.Uint64, shape = new [] { x.Length } },
-            bool[]          x => new() { data = await Upload(x.ToStream()), type = Dtype.Bool, shape = new [] { x.Length } },
-            Tensor<float>   x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Float32, shape = x.shape },
-            Tensor<double>  x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Float64, shape = x.shape },
-            Tensor<sbyte>   x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Int8, shape = x.shape },
-            Tensor<short>   x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Int16, shape = x.shape },
-            Tensor<int>     x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Int32, shape = x.shape },
-            Tensor<long>    x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Int64, shape = x.shape },
-            Tensor<byte>    x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Uint8, shape = x.shape },
-            Tensor<ushort>  x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Uint16, shape = x.shape },
-            Tensor<uint>    x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Uint32, shape = x.shape },
-            Tensor<ulong>   x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Uint64, shape = x.shape },
-            Tensor<bool>    x => new() { data = await Upload(x.data.ToStream()), type = Dtype.Bool, shape = x.shape },
-            string          x => new() { data = await Upload(x.ToStream(), mime: @"text/plain"), type = Dtype.String },
-            IList           x => new() { data = await Upload(JsonConvert.SerializeObject(x).ToStream(), mime: @"application/json"), type = Dtype.List },
-            IDictionary     x => new() { data = await Upload(JsonConvert.SerializeObject(x).ToStream(), mime: @"application/json"), type = Dtype.Dict },
-            Image           x => new() { data = "", type = Dtype.Image },
-            Stream          x => new() { data = await Upload(x), type = Dtype.Binary },
+            null              => new() { dtype = Dtype.Null },
+            float           x => new() { data = await Upload(new [] { x }.ToStream()), dtype = Dtype.Float32, shape = new int[0] },
+            double          x => new() { data = await Upload(new [] { x }.ToStream()), dtype = Dtype.Float64, shape = new int[0] },
+            sbyte           x => new() { data = await Upload(new [] { x }.ToStream()), dtype = Dtype.Int8, shape = new int[0] },
+            short           x => new() { data = await Upload(new [] { x }.ToStream()), dtype = Dtype.Int16, shape = new int[0] },
+            int             x => new() { data = await Upload(new [] { x }.ToStream()), dtype = Dtype.Int32, shape = new int[0] },
+            long            x => new() { data = await Upload(new [] { x }.ToStream()), dtype = Dtype.Int64, shape = new int[0] },
+            byte            x => new() { data = await Upload(new [] { x }.ToStream()), dtype = Dtype.Uint8, shape = new int[0] },
+            ushort          x => new() { data = await Upload(new [] { x }.ToStream()), dtype = Dtype.Uint16, shape = new int[0] },
+            uint            x => new() { data = await Upload(new [] { x }.ToStream()), dtype = Dtype.Uint32, shape = new int[0] },
+            ulong           x => new() { data = await Upload(new [] { x }.ToStream()), dtype = Dtype.Uint64, shape = new int[0] },
+            bool            x => new() { data = await Upload(new [] { x }.ToStream()), dtype = Dtype.Bool, shape = new int[0] },
+            float[]         x => new() { data = await Upload(x.ToStream()), dtype = Dtype.Float32, shape = new [] { x.Length } },
+            double[]        x => new() { data = await Upload(x.ToStream()), dtype = Dtype.Float64, shape = new [] { x.Length } },
+            sbyte[]         x => new() { data = await Upload(x.ToStream()), dtype = Dtype.Int8, shape = new [] { x.Length } },
+            short[]         x => new() { data = await Upload(x.ToStream()), dtype = Dtype.Int16, shape = new [] { x.Length } },
+            int[]           x => new() { data = await Upload(x.ToStream()), dtype = Dtype.Int32, shape = new [] { x.Length } },
+            long[]          x => new() { data = await Upload(x.ToStream()), dtype = Dtype.Int64, shape = new [] { x.Length } },
+            byte[]          x => new() { data = await Upload(x.ToStream()), dtype = Dtype.Uint8, shape = new [] { x.Length } },
+            ushort[]        x => new() { data = await Upload(x.ToStream()), dtype = Dtype.Uint16, shape = new [] { x.Length } },
+            uint[]          x => new() { data = await Upload(x.ToStream()), dtype = Dtype.Uint32, shape = new [] { x.Length } },
+            ulong[]         x => new() { data = await Upload(x.ToStream()), dtype = Dtype.Uint64, shape = new [] { x.Length } },
+            bool[]          x => new() { data = await Upload(x.ToStream()), dtype = Dtype.Bool, shape = new [] { x.Length } },
+            Tensor<float>   x => new() { data = await Upload(x.data.ToStream()), dtype = Dtype.Float32, shape = x.shape },
+            Tensor<double>  x => new() { data = await Upload(x.data.ToStream()), dtype = Dtype.Float64, shape = x.shape },
+            Tensor<sbyte>   x => new() { data = await Upload(x.data.ToStream()), dtype = Dtype.Int8, shape = x.shape },
+            Tensor<short>   x => new() { data = await Upload(x.data.ToStream()), dtype = Dtype.Int16, shape = x.shape },
+            Tensor<int>     x => new() { data = await Upload(x.data.ToStream()), dtype = Dtype.Int32, shape = x.shape },
+            Tensor<long>    x => new() { data = await Upload(x.data.ToStream()), dtype = Dtype.Int64, shape = x.shape },
+            Tensor<byte>    x => new() { data = await Upload(x.data.ToStream()), dtype = Dtype.Uint8, shape = x.shape },
+            Tensor<ushort>  x => new() { data = await Upload(x.data.ToStream()), dtype = Dtype.Uint16, shape = x.shape },
+            Tensor<uint>    x => new() { data = await Upload(x.data.ToStream()), dtype = Dtype.Uint32, shape = x.shape },
+            Tensor<ulong>   x => new() { data = await Upload(x.data.ToStream()), dtype = Dtype.Uint64, shape = x.shape },
+            Tensor<bool>    x => new() { data = await Upload(x.data.ToStream()), dtype = Dtype.Bool, shape = x.shape },
+            string          x => new() { data = await Upload(x.ToStream(), mime: @"text/plain"), dtype = Dtype.String },
+            IList           x => new() { data = await Upload(JsonConvert.SerializeObject(x).ToStream(), mime: @"application/json"), dtype = Dtype.List },
+            IDictionary     x => new() { data = await Upload(JsonConvert.SerializeObject(x).ToStream(), mime: @"application/json"), dtype = Dtype.Dict },
+            Image           x => new() { data = "", dtype = Dtype.Image },
+            Stream          x => new() { data = await Upload(x), dtype = Dtype.Binary },
             Enum            x => await ToValue(x.ToObject()),
             RemoteValue     x => x,
             _                 => throw new InvalidOperationException($"Failed to serialize value '{value}' of type `{value.GetType()}` because it is not supported"),
         };
 
-        private async Task<object?> ToObject(RemoteValue value) { // INCOMPLETE // Image
-            if (value.type == Dtype.Null)
+        private async Task<object?> ToObject(RemoteValue value) {
+            if (value.dtype == Dtype.Null)
                 return null;
             using var stream = await Download(value.data!);
-            return value.type switch {
+            return value.dtype switch {
                 Dtype.Float32   => stream.ToObject<float>(value.shape!),
                 Dtype.Float64   => stream.ToObject<double>(value.shape!),
                 Dtype.Int8      => stream.ToObject<sbyte>(value.shape!),
@@ -138,9 +138,9 @@ namespace Muna.Beta.Services {
                 Dtype.String    => new StreamReader(stream).ReadToEnd(),
                 Dtype.List      => JsonConvert.DeserializeObject<JArray>(new StreamReader(stream).ReadToEnd()),
                 Dtype.Dict      => JsonConvert.DeserializeObject<JObject>(new StreamReader(stream).ReadToEnd()),
-                Dtype.Image     => default,
+                Dtype.Image     => DeserializeImageValue(stream),
                 Dtype.Binary    => stream.Clone(),
-                _               => throw new InvalidOperationException($"Failed to deserialize value with type {value.type} because it is not supported"),
+                _               => throw new InvalidOperationException($"Failed to deserialize value with type {value.dtype} because it is not supported"),
             };
         }
 
@@ -161,6 +161,11 @@ namespace Muna.Beta.Services {
                 return new MemoryStream(data, 0, data.Length, false, false);
             }
             return await client.Download(url);
+        }
+
+        private static Image DeserializeImageValue(Stream stream) {
+            using var value = C.Value.CreateFromBinary(stream, @"image/*");
+            return (Image)value.ToObject()!;
         }
 
         [Preserve, Serializable]