+++ Changelog.md
@@ -1,5 +1,8 @@
 ## 0.0.50
++ Added support for `aac`, `flac`, `mp3`, `opus`, `wav` response formats in `muna.Beta.Openai.Audio.Speech.Create` method.
++ Fixed errors when making remote predictions with image inputs or outputs.
 + Fixed sporadic timeout error when making some predictions.
++ Replaced `Parameter.range` field to `Parameter.min` and `Parameter.max` fields.
 
 ## 0.0.49
 + Fixed sporadic segmentation fault when making predictions on input tensors.